requirements.txt
hvac~=1.0
PyYaml~=6.0